personal_code_space_additions/requirements.txt
python-dotenv
psycopg2-binary
#semantic-kernel==1.23.1
azure-identity
requests
ipykernel
openai
langchain
langgraph
pandas
python-dotenv
langtrace-python-sdk
botocore
langchain_community
langchain_experimental
matplotlib
autogen
autogen-agentchat~=0.2
seaborn
azure-ai-inference
azure-identity
azure-ai-projects
semantic-kernel==1.27.1
semantic-kernel[mcp]
azure-ai-ml
google-api-python-client
opentelemetry-exporter-otlp-proto-grpc
pyodbc
bs4